requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Werkzeug==2.2.3
psycopg2-binary==2.9.6
mysql-connector-python==8.0.33
prometheus-client==0.17.0
psutil==5.9.5
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0
cryptography==41.0.3
email-validator==2.0.0.post2

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0
